Refactor financial and market regime modules
